Migrations Fixed
Troubles with index solved.

+++ database/migrations/2020_03_29_222852_add_rol_id_to_users.php
@@ -14,7 +14,7 @@ class AddRolIdToUsers extends Migration
     public function up()
     {
         Schema::table('users', function (Blueprint $table) {
-            $table->unsignedInteger('rol_id')->nullable(); 
+            $table->unsignedBigInteger('rol_id')->nullable(); 
             $table->foreign('rol_id')->references('id')->on('rols');
         });
     }
